pyproject.toml
[tool.poetry]
name = "screenwriter"
version = "0.0.3"
description = "ScreenWriter is a simple util that makes your daily life at university easier."
authors = ["CrazyProger1 <crazyproger1@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
translatable-enums = "^0.0.3"
python-docx = "^1.1.0"
pydantic = "^2.6.3"
typeguard = "^4.1.5"
toml = "^0.10.2"
art = "^6.1"
colorama = "^0.4.6"
keyboard = "^0.13.5"
pyperclip = "^1.8.2"
pillow = "^10.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
colorlog = "^6.8.2"
pyinstaller = "^6.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
